Move Mock in External Project
