+++ src/PolarisDesk.API/Services/CustomerService.cs
@@ -1,5 +1,5 @@
-﻿using PolarisDesk.API.Interface;
-using PolarisDesk.Models;
+﻿using PolarisDesk.Models;
+using PolarisDesk.Shared.Interface;
 using System;
 using System.Threading.Tasks;
