+++ src/PolarisDesk.API/Services/TicketStatusService.cs
@@ -1,5 +1,6 @@
-﻿using PolarisDesk.API.Interface;
+﻿
 using PolarisDesk.Models;
+using PolarisDesk.Shared.Interface;
 using System;
 using System.Threading.Tasks;
