+++ src/PolarisDesk.API/Startup.cs
@@ -5,9 +5,11 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using PolarisDesk.API.Interface;
+
 using PolarisDesk.API.Services;
 using PolarisDesk.Models;
+using PolarisDesk.Shared.Interface;
+using PolarisDesk.Test.Services;
 using System;
 
 namespace PolarisDesk.API
