+++ src/PolarisDesk.Test/ServicesMock/TicketServiceMock.cs
@@ -1,13 +1,14 @@
 ﻿
 using Bogus;
-using PolarisDesk.API.Interface;
+
 using PolarisDesk.Models;
+using PolarisDesk.Shared.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PolarisDesk.API.Services
+namespace PolarisDesk.Test.Services
 {
 	public class TicketServiceMock : ICrudService<Ticket, Guid>
 	{
@@ -34,7 +35,7 @@ namespace PolarisDesk.API.Services
 		public async Task Delete(Guid id)
 		{
 			var ticket = await Get(id);
-			if (ticket is not null)
+			if (ticket != null)
 			{
 				ticket.Enabled = false;
 				ticket.Deleted = DateTime.Now;
@@ -54,7 +55,7 @@ namespace PolarisDesk.API.Services
 		public async Task Update(Ticket item)
 		{
 			var ticket = await Get(item.ID);
-			if (ticket is not null)
+			if (ticket != null)
 			{
 				ticket.Title = item.Title;
 				ticket.Code = item.Code;
