+++ src/PolarisDesk.Test/ServicesMock/TicketStatusServiceMock.cs
@@ -1,12 +1,12 @@
 ﻿using Bogus;
-using PolarisDesk.API.Interface;
 using PolarisDesk.Models;
+using PolarisDesk.Shared.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PolarisDesk.API.Services
+namespace PolarisDesk.Test.Services
 {
 	public class TicketStatusServiceMock : ICrudService<TicketStatus, Guid>
 	{
@@ -31,7 +31,7 @@ namespace PolarisDesk.API.Services
 		public async Task Delete(Guid id)
 		{
 			var ticket = await Get(id);
-			if (ticket is not null)
+			if (ticket != null)
 			{
 				ticket.Enabled = false;
 				ticket.Deleted = DateTime.Now;
@@ -51,7 +51,7 @@ namespace PolarisDesk.API.Services
 		public async Task Update(TicketStatus item)
 		{
 			var ticket = await Get(item.ID);
-			if (ticket is not null)
+			if (ticket != null)
 			{
 				ticket.Name = item.Name;
 				ticket.Updated = DateTime.Now;
