+++ src/PolarisDesk.sln
@@ -13,6 +13,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PolarisDesk.API", "PolarisD
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PolarisDesk.Models", "PolarisDesk.Models\PolarisDesk.Models.csproj", "{8913C315-060B-433C-B388-1054920EC79C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PolarisDesk.Test", "PolarisDesk.Test\PolarisDesk.Test.csproj", "{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -83,6 +85,18 @@ Global
 		{8913C315-060B-433C-B388-1054920EC79C}.Release|x64.Build.0 = Release|Any CPU
 		{8913C315-060B-433C-B388-1054920EC79C}.Release|x86.ActiveCfg = Release|Any CPU
 		{8913C315-060B-433C-B388-1054920EC79C}.Release|x86.Build.0 = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|x64.Build.0 = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Debug|x86.Build.0 = Debug|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|x64.ActiveCfg = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|x64.Build.0 = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|x86.ActiveCfg = Release|Any CPU
+		{CEC6F0CD-477F-44D8-BC06-2C360A696C4E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
